32bit fixes

+++ src/unittest/test_partitions.cxx
@@ -2,11 +2,12 @@
 #include <iostream>
 #include "opengm/utilities/partitions.hxx"
 #include <opengm/unittests/test.hxx>
+#include <stdint.h>
 
 
 void testPartition()
 {
-   opengm::Partitions<size_t,size_t> P;
+   opengm::Partitions<uint64_t,uint64_t> P;
    P.resize(3); 
    std::cout <<"Test BellNumber: ..."<<std::flush;
    OPENGM_TEST(P.BellNumber(2)==2);
@@ -20,8 +21,8 @@ void testPartition()
    OPENGM_TEST(P.getPartition(4)==7); 
    OPENGM_TEST(P.getPartition(5)>8);
 
-   std::vector<size_t> ltest(3);
-   for(size_t i=0; i<5; ++i){
+   std::vector<uint64_t> ltest(3);
+   for(uint64_t i=0; i<5; ++i){
       //const size_t el = P.getPartition(i);
       P.getPartition(i,ltest);
       //std::cout << P.label2Index(ltest) <<" ("<< el <<") "<<ltest[0]<<ltest[1]<<ltest[2]<<std::endl;
@@ -33,16 +34,16 @@ void testPartition()
 
    std::cout <<"Test label2Index: ..."<<std::flush;
    P.resize(3);
-   std::vector<size_t> l(3,0);
+   std::vector<uint64_t> l(3,0);
    OPENGM_TEST(P.label2Index(l)==4);
    l[0]=1;
    OPENGM_TEST(P.label2Index(l)==3);
    l[1]=2;
    OPENGM_TEST(P.label2Index(l)==0); 
 
-   for(size_t b=4;b<12;++b){
+   for(uint64_t b=4;b<12;++b){
       P.resize(b);
-      std::vector<size_t> l(b,0);
+      std::vector<uint64_t> l(b,0);
       OPENGM_TEST(P.label2Index(l)==P.BellNumber(b)-1);
    }
    std::cout <<"  OK!"<<std::endl;
